+++ src/gen/genModule.ts
@@ -17,6 +17,8 @@ export const GEN_FILE_PATTERN = SEPARATOR + GEN_FILE_PATTERN_TEXT + SEPARATOR;
 
 export const GEN_SCHEMA_FILE_PATTERN_TEXT = 'schema';
 export const GEN_SCHEMA_FILE_PATTERN = SEPARATOR + GEN_SCHEMA_FILE_PATTERN_TEXT + SEPARATOR;
+export const GEN_SCHEMA_PATH_SUFFIX = GEN_SCHEMA_FILE_PATTERN + 'ts';
+export const GEN_SCHEMA_IDENTIFIER_SUFFIX = 'Schema';
 
 export type GenModuleType = 'basic' | 'schema';
 export type GenModule = BasicGenModule | SchemaGenModule;