+++ src/utils/schema.ts
@@ -1,7 +1,9 @@
+import {traverse} from '@feltjs/util/object.js';
 import type {JSONSchema} from '@ryanatkn/json-schema-to-typescript';
 import type {ResolverOptions} from 'json-schema-ref-parser';
 import type z from 'zod';
 import {zodToJsonSchema} from 'zod-to-json-schema';
+import type {GenContext} from '../gen/gen';
 
 export interface VocabSchema extends JSONSchema {
 	$id: string;
@@ -31,3 +33,35 @@ export const toVocabSchemaResolver = (schemas: VocabSchema[]): ResolverOptions =
 		return JSON.stringify(schema);
 	},
 });
+
+// TODO do this more robustly (handle `/`?)
+const parseSchemaName = ($id: string): string | null =>
+	$id.startsWith('/schemas/') && $id.endsWith('.json') ? $id.substring(9, $id.length - 5) : null;
+
+// TODO make an option, is very hardcoded
+const toSchemaImport = ($id: string, ctx: GenContext): string | null => {
+	if (!$id.startsWith('/schemas/') || !$id.endsWith('.json')) return null;
+	const name = $id.substring(9, $id.length - 5);
+	return name in ctx.imports ? ctx.imports[name] : null;
+};
+
+/**
+ * Mutates `schema` with `tsType` and `tsImport`, if appropriate.
+ * @param schema
+ */
+export const inferSchemaTypes = (schema: VocabSchema, ctx: GenContext): void => {
+	traverse(schema, (key, value, obj) => {
+		if (key === '$ref') {
+			if (!('tsType' in obj)) {
+				const tsType = parseSchemaName(value);
+				if (tsType) obj.tsType = tsType;
+			}
+			if (!('tsImport' in obj)) {
+				const tsImport = toSchemaImport(value, ctx);
+				if (tsImport) obj.tsImport = tsImport;
+			}
+		} else if (key === 'instanceof') {
+			if (!('tsType' in obj)) obj.tsType = value;
+		}
+	});
+};