+++ changelog.md
@@ -1,5 +1,12 @@
 # changelog
 
+## 0.75.0
+
+- **break**: change schema type generation to infer `tsType` and `tsImport` from `$ref` and `instanceof`
+  ([#370](https://github.com/feltjs/gro/pull/370))
+- **break**: bump peer dep `svelte@0.58.0`
+  ([#370](https://github.com/feltjs/gro/pull/370))
+
 ## 0.74.0
 
 - **break**: fix schema type generation to be shallow, excluding the types of referenced schemas